Create a huge instance to experiment with

+++ lib/creater.ts
@@ -2,12 +2,13 @@ import cdk = require('@aws-cdk/core');
 import ec2 = require('@aws-cdk/aws-ec2');
 import autoscaling = require('@aws-cdk/aws-autoscaling');
 import { Vpc, AmazonLinuxImage, AmazonLinuxGeneration } from '@aws-cdk/aws-ec2';
-import { BlockDeviceVolume, UpdateType } from '@aws-cdk/aws-autoscaling';
+import { BlockDeviceVolume, UpdateType, AutoScalingGroup } from '@aws-cdk/aws-autoscaling';
 
 export interface CraterProps {
     vpc: ec2.IVpc,
     instanceType: string,
     keyName: string,
+    ebs: boolean,
 }
 
 export class Crater extends cdk.Construct {
@@ -24,31 +25,37 @@ export class Crater extends cdk.Construct {
             machineImage: new AmazonLinuxImage({
                 generation: AmazonLinuxGeneration.AMAZON_LINUX_2,
             }),
-            blockDevices: [{
+            blockDevices: props.ebs ? [{
                 deviceName: "/dev/sdf",
                 volume: BlockDeviceVolume.ebs(2000),
-            }],
+            }] : []
         })
 
         // Allow SSH from anywhere
         asg.connections.allowFromAnyIpv4(ec2.Port.tcp(22));
 
+
+        // If we're using an additional EBS volume, make it available
+        // for both crater (/root) and docker (/var/lib/docker)
+        if (props.ebs) {
+            asg.addUserData("mkfs -t xfs /dev/nvme1n1");
+            asg.addUserData("mount /dev/nvme1n1 /root");
+            asg.addUserData("mkdir -p /root/docker");
+            asg.addUserData("ln -sf /root/docker /var/lib/docker");
+        }
+
         // Set up crater
-        asg.addUserData("mkfs -t xfs /dev/nvme1n1");
-        asg.addUserData("mount /dev/nvme1n1 /root");
-        asg.addUserData("mkdir -p /root/docker");
-        asg.addUserData("ln -sf /root/docker /var/lib/docker");
         asg.addUserData("yum update -y");
-        asg.addUserData("yum groupinstall 'Development Tools'");
-        asg.addUserData("yum install -y git docker sqlite-devel openssl-devel");
+        asg.addUserData("yum groupinstall -y 'Development Tools'");
+        asg.addUserData("yum install -y git docker tmux htop sqlite-devel openssl-devel");
         asg.addUserData("systemctl enable docker.service");
         asg.addUserData("systemctl start docker.service");
         asg.addUserData("docker pull rustops/crates-build-env");
-        asg.addUserData("git clone https://github.com/rust-lang/crater");
+        asg.addUserData("git clone https://github.com/rust-lang/crater /root/crater");
         asg.addUserData("curl https://static.rust-lang.org/rustup/dist/x86_64-unknown-linux-gnu/rustup-init >/tmp/rustup-init");
         asg.addUserData("chmod +x /tmp/rustup-init");
         asg.addUserData("/tmp/rustup-init -y --no-modify-path --default-toolchain stable --profile minimal");
-        asg.addUserData("export PATH=/root/.cargo/bin:$PATH")
+        asg.addUserData("echo 'export PATH=/root/.cargo/bin:$PATH' >> /root/.bashrc")
 
     }
